Restored Import functionality for exported games
On UserInitialize, check to see if there is a project to be imported.

Added a new StartupPhaseResult enumeration. This can be used to determine if the next initialization phase should be run.

Load shortcuts and documentation after user initializiation.
The MainConfig data directory where they are loaded from can be modified by UserInitialize (e.g.: for imported games)